backend/pyproject.toml
[tool.poetry]
name = "job-agent-backend"
version = "0.1.0"
description = "AI-powered job application assistant backend"
authors = []
packages = [{include = "backend"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115"
uvicorn = {extras = ["standard"], version = "^0.29"}
anthropic = "^0.92"
pdfplumber = "^0.11"
python-docx = "^1.1"
sqlalchemy = "^2.0"
aiosqlite = "^0.20"
sse-starlette = "^2.1"
python-multipart = "^0.0.9"
python-dotenv = "^1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.23"
httpx = "^0.27"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
